move implementation from headerfiles to dedicated cpp files
