+++ platformio.ini
@@ -13,19 +13,24 @@ platform = espressif32
 board = esp32dev
 framework = arduino
 lib_deps = 
-	adafruit/Adafruit BusIO@^1.16.2
 	adafruit/Adafruit PWM Servo Driver Library@^3.0.2
 	fmalpartida/LiquidCrystal@^1.5.0
 	mbed-shaga/AdafruitPwmServoDriver@0.0.0+sha.0b34adf7d259
 	tomstewart89/BasicLinearAlgebra@^5.1
+	paulstoffregen/Encoder@^1.4.4
+	adafruit/Adafruit BusIO@^1.16.2
 
 [env:debugging]
 platform = espressif32
 board = esp32dev
 framework = arduino
-build_src_filter = +<webserver.cpp>
+build_src_filter = +<inverse_kinematics.cpp>
 lib_deps = 
 	adafruit/Adafruit PWM Servo Driver Library@^3.0.2
 	fmalpartida/LiquidCrystal@^1.5.0
 	mbed-shaga/AdafruitPwmServoDriver@0.0.0+sha.0b34adf7d259
 	tomstewart89/BasicLinearAlgebra@^5.1
+	paulstoffregen/Encoder@^1.4.4
+	adafruit/Adafruit BusIO@^1.16.2
+	Wire
+	SPI
